requirements.txt
accelerate==1.10.1
certifi==2025.6.15
charset-normalizer==3.4.2
chronos-forecasting==1.5.3
contourpy==1.3.2
cycler==0.12.1
einops==0.8.1
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
hf-xet==1.1.5
huggingface-hub==0.33.0
idna==3.10
Jinja2==3.1.3
joblib==1.5.1
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==25.0
pandas==2.3.0
pillow==11.2.1
psutil==7.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
sympy==1.13.3
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.1.0+cu121
torchaudio==2.1.0+cu121
torchvision==0.16.0+cu121
tqdm==4.67.1
transformers==4.52.4
triton==2.1.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
